Change component structure
